Create interfaces in models folders and update paths
