+++ src/main/webapp/src/app/models/geometry.ts
@@ -1,0 +1,9 @@
+import { Location } from './location';
+
+export interface Geometry {
+  location: Location;
+  viewport: {
+    northeast: Location;
+    southwest: Location;
+  }
+}
