+++ src/main/webapp/src/app/models/hidden-gem.ts
@@ -1,0 +1,18 @@
+import { Geometry } from './geometry';
+import { Photos } from './photos';
+import { OpeningHours } from './openingHours';
+
+export interface HiddenGem {
+  geometry: Geometry;
+  name: string;
+  icon: string;
+  placeId: string;
+  rating: number;
+  types: string[];
+  openingHours: OpeningHours;
+  photos: Photos[];
+  vicinity: string;
+  permanentlyClosed: boolean;
+  userRatingsTotal: number;
+  businessStatus: string;
+}
