+++ src/main/webapp/src/app/models/openingHours.ts
@@ -1,0 +1,3 @@
+export interface OpeningHours {
+  openNow: boolean;
+}
