+++ src/main/webapp/src/app/models/photos.ts
@@ -1,0 +1,6 @@
+export interface Photos {
+  photoReference: string;
+  height: number;
+  width: number;
+  htmlAttributions: string[];
+}
